pyproject.toml
[project]
name = "tantivy-search"
version = "0.1.0"
description = "Local code and markdown search tool built on Tantivy"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Tilo Himmelsbach", email = "dertilo@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Indexing",
]
dependencies = [
    "tantivy>=0.25.0",
    "python-frontmatter>=1.1.0",
    "langchain-text-splitters>=0.3.0",
]

[project.optional-dependencies]
mcp = ["fastmcp>=2.0.0"]

[project.urls]
Repository = "https://github.com/dertilo/tantivy-search"

[project.scripts]
tantivy-search = "tantivy_search.cli:main"
tantivy_search = "tantivy_search.cli:main"
tantivy-index = "tantivy_search.cli:main_index"
tantivy_index = "tantivy_search.cli:main_index"
tantivy-search-mcp = "tantivy_search.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tantivy_search"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
addopts = "-v"
